extract useAppScrollParent hook; resolve scroll parent before first paint
Second pass on the virtuoso migration. Two concrete fixes to the
'take time to load, ghost empty space, rows appear only on first
scroll' symptoms:

- Extract the DOM-walk-to-find-scroll-parent logic into
  useAppScrollParent (hooks/useAppScrollParent.js). Runs in
  useLayoutEffect so the scroll parent is resolved synchronously
  after mount and before the browser paints, not in a useEffect
  that fires after paint.
- Both list sections render null until the hook reports ready.
  Without the gate, virtuoso mounted once with useWindowScroll=true,
  then swapped to customScrollParent after the useEffect committed —
  that swap left stale internal state and caused the 'rows invisible
  until first scroll' flicker.
- PrincipleDetailPage calls the hook once at the page level and
  passes scrollParent down to both ViolationListSection and
  ComplianceListSection, so both sections share one scrollbar. Also
  drops the residual showAllCompliance / PAGE_SIZE plumbing — dead
  code once virtuoso handles lazy rendering.
- HistoryPage uses the same hook; unblocks the same 'ghost empty'
  regression on the evaluations list.
